run virsh net-undefine; load br_netfilter on RHEL 7.3

+++ recipes/eucanetd.rb
@@ -4,11 +4,24 @@ require 'chef/version_constraint'
 include_recipe "eucalyptus::default"
 
 # Remove default virsh network which runs its own dhcp server
-execute 'virsh net-destroy default' do
-  ignore_failure true
+if Chef::VersionConstraint.new("~> 6.0").include?(node['platform_version'])
+  execute 'virsh net-destroy default' do
+    ignore_failure true
+  end
+  execute 'virsh net-autostart default --disable' do
+    ignore_failure true
+  end
 end
-execute 'virsh net-autostart default --disable' do
-  ignore_failure true
+if Chef::VersionConstraint.new("~> 7.0").include?(node['platform_version'])
+  execute 'virsh net-destroy default' do
+    ignore_failure true
+  end
+  execute 'virsh net-autostart default --disable' do
+    ignore_failure true
+  end
+  execute 'virsh net-undefine default' do
+    ignore_failure true
+  end
 end
 
 if node["eucalyptus"]["install-type"] == "packages"
